更改QQ号为Long存储，避免溢出

+++ message_service_1/src/main/java/org/xzx/bean/messageBean/Send_Message/WebSocketMessage.java
@@ -12,7 +12,7 @@ public class WebSocketMessage {
 
     private String sender;
 
-    private int sender_id;
+    private long sender_id;
 
     private int isImg;
